Made changes to tables related to forms.
Now foreign key constraints have a name. Question related tables now cascade on delete and on update. There was a problem with the model with some foreign keys to Grupo that is now fixed.

+++ AppIntegrador/Models/Periodo_activa_por.cs
@@ -16,9 +16,9 @@ namespace AppIntegrador.Models
     {
         public string FCodigo { get; set; }
         public string CSigla { get; set; }
-        public int GNumero { get; set; }
+        public byte GNumero { get; set; }
         public int GAnno { get; set; }
-        public int GSemestre { get; set; }
+        public byte GSemestre { get; set; }
         public System.DateTime FechaInicio { get; set; }
         public System.DateTime FechaFin { get; set; }
